Making it clearer where doxygen runs
Took me a while to troubleshoot this, but finally figure it out. I made it a bit clearer that `doxygen` is ran from the root directory of the repo. We can leave it as is and accept this change, or consider running `doxygen` where the user has its Doxyfile, maybe have a `doxygenconf`: name of the Doxyfile, and doxygenconfpath: path of this file. This would save the user the work of changing every path it might have in its doxyfile if he set it up as to run doxygen from the file's location